clean up loss related types
