+++ src/deepqmc/loss/__init__.py
@@ -5,11 +5,14 @@ from .clip import (
     median_log_squeeze_and_mask,
     psi_ratio_clip_and_mask,
 )
-from .loss_function import LossFunctionFactory, create_loss_fn
+from .loss_function import create_loss_fn
+from .base import LossAndGradFunction, LossFunction, LossFunctionFactory
 
 __all__ = [
     'LocalEnergyClipAndMaskFn',
     'LossFunctionFactory',
+    'LossAndGradFunction',
+    'LossFunction',
     'PsiRatioClipAndMaskFn',
     'median_clip_and_mask',
     'median_log_squeeze_and_mask',
